Revert "Add app infrastructure for component branch merge"
This reverts commit 93495e85db0080c659ced2dc03257d66fe9915b3.
