+++ package.json
@@ -10,13 +10,10 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^1.13.1",
     "lucide-react": "^1.7.0",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
-    "react-icons": "^5.6.0",
-    "react-router-dom": "^7.9.5",
-    "zustand": "^5.0.8"
+    "react-icons": "^5.6.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.39.4",
