--- public/manifest.webmanifest
@@ -1,25 +1,0 @@
-{
-  "name": "SOLve",
-  "short_name": "SOLve",
-  "description": "ESG 활동 기반 대안 신용평가 플랫폼",
-  "start_url": "/home",
-  "scope": "/",
-  "display": "standalone",
-  "orientation": "portrait",
-  "background_color": "#F1F5F9",
-  "theme_color": "#0046FF",
-  "icons": [
-    {
-      "src": "/pwa-icon.svg",
-      "sizes": "any",
-      "type": "image/svg+xml",
-      "purpose": "any"
-    },
-    {
-      "src": "/pwa-maskable.svg",
-      "sizes": "any",
-      "type": "image/svg+xml",
-      "purpose": "maskable"
-    }
-  ]
-}
