--- public/sw.js
@@ -1,48 +1,0 @@
-const STATIC_CACHE = 'solve-static-v1'
-const APP_SHELL = ['/', '/index.html', '/manifest.webmanifest', '/pwa-icon.svg', '/pwa-maskable.svg']
-
-self.addEventListener('install', (event) => {
-  event.waitUntil(
-    caches.open(STATIC_CACHE).then((cache) => cache.addAll(APP_SHELL)),
-  )
-  self.skipWaiting()
-})
-
-self.addEventListener('activate', (event) => {
-  event.waitUntil(
-    caches.keys().then((keys) =>
-      Promise.all(
-        keys
-          .filter((key) => key !== STATIC_CACHE)
-          .map((key) => caches.delete(key)),
-      ),
-    ),
-  )
-  self.clients.claim()
-})
-
-self.addEventListener('fetch', (event) => {
-  if (event.request.method !== 'GET') {
-    return
-  }
-
-  const requestUrl = new URL(event.request.url)
-
-  if (requestUrl.origin !== self.location.origin) {
-    return
-  }
-
-  if (event.request.mode === 'navigate') {
-    event.respondWith(
-      fetch(event.request).catch(async () => {
-        const cache = await caches.open(STATIC_CACHE)
-        return (await cache.match('/index.html')) || Response.error()
-      }),
-    )
-    return
-  }
-
-  event.respondWith(
-    caches.match(event.request).then((cachedResponse) => cachedResponse || fetch(event.request)),
-  )
-})
