--- src/constants/routePaths.ts
@@ -1,15 +1,0 @@
-export const ROUTE_PATHS = {
-  root: '/',
-  login: '/login',
-  signup: '/signup',
-  onboarding: '/onboarding',
-  home: '/home',
-  activities: '/activities',
-  activityEnvironment: '/activities/environment',
-  activitySocial: '/activities/social',
-  activityGovernance: '/activities/governance',
-  finance: '/finance',
-  my: '/my',
-  chatbot: '/chatbot',
-  componentPreview: '/preview/components',
-} as const
