--- src/pages/activities/EnvironmentPage.tsx
@@ -1,11 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function EnvironmentPage() {
-  return (
-    <PageScaffold
-      title="환경 활동"
-      description="OCR 인증 기반 E 활동 페이지 자리입니다."
-      currentTab="benefits"
-    />
-  )
-}
