--- src/pages/activities/GovernancePage.tsx
@@ -1,11 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function GovernancePage() {
-  return (
-    <PageScaffold
-      title="자기관리 활동"
-      description="출석, 퀴즈, 금융 습관형 활동이 들어갈 자리입니다."
-      currentTab="benefits"
-    />
-  )
-}
