--- src/pages/activities/SocialPage.tsx
@@ -1,11 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function SocialPage() {
-  return (
-    <PageScaffold
-      title="사회 활동"
-      description="기부, 가치가게, 봉사 흐름이 연결될 자리입니다."
-      currentTab="benefits"
-    />
-  )
-}
