--- src/pages/auth/LoginPage.tsx
@@ -1,10 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function LoginPage() {
-  return (
-    <PageScaffold
-      title="로그인"
-      description="실제 로그인 API가 연결되기 전까지 사용하는 인증 진입 페이지입니다."
-    />
-  )
-}
