--- src/pages/auth/OnboardingPage.tsx
@@ -1,10 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function OnboardingPage() {
-  return (
-    <PageScaffold
-      title="온보딩 설문"
-      description="사용자 유형 분류 설문과 첫 진입 플로우를 연결할 자리입니다."
-    />
-  )
-}
