--- src/pages/auth/SignupPage.tsx
@@ -1,10 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function SignupPage() {
-  return (
-    <PageScaffold
-      title="회원가입"
-      description="백엔드 스펙이 정리되면 여기서 실제 회원가입 흐름을 연결합니다."
-    />
-  )
-}
