--- src/pages/chatbot/ChatbotPage.tsx
@@ -1,11 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function ChatbotPage() {
-  return (
-    <PageScaffold
-      title="AI 챗봇"
-      description="추천 결과와 사용자 상태를 재사용하는 챗봇 화면 자리입니다."
-      currentTab="home"
-    />
-  )
-}
