--- src/pages/finance/FinancePage.tsx
@@ -1,11 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function FinancePage() {
-  return (
-    <PageScaffold
-      title="금융 상품"
-      description="적금과 대출 추천, 우대금리 정보를 연결할 자리입니다."
-      currentTab="finance"
-    />
-  )
-}
