--- src/pages/home/HomePage.tsx
@@ -1,11 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function HomePage() {
-  return (
-    <PageScaffold
-      title="메인 대시보드"
-      description="등급, 포인트, AI 추천, 금융 추천을 연결하는 메인 화면 자리입니다."
-      currentTab="home"
-    />
-  )
-}
