--- src/pages/my/MyPage.tsx
@@ -1,11 +1,0 @@
-import { PageScaffold } from '../PageScaffold'
-
-export function MyPage() {
-  return (
-    <PageScaffold
-      title="마이페이지"
-      description="내 정보, 활동 이력, 포인트/등급 히스토리를 연결할 자리입니다."
-      currentTab="mypage"
-    />
-  )
-}
