--- src/store/index.ts
@@ -1,2 +1,0 @@
-export { useAppStore } from './appStore'
-export { useAuthStore } from './authStore'
