--- src/styles/globals.css
@@ -1,18 +1,0 @@
-:root {
-  font-family: var(--font-pretendard);
-  color: var(--color-font-main);
-  background: var(--color-bg-light);
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-body {
-  min-width: 320px;
-  background: linear-gradient(180deg, #f8fafc 0%, #f1f5f9 100%);
-}
-
-#root {
-  min-height: 100vh;
-}
