--- src/types/api.ts
@@ -1,6 +1,0 @@
-export interface ApiErrorPayload {
-  status: number
-  message: string
-  code?: string
-  details?: unknown
-}
