--- src/types/auth.ts
@@ -1,6 +1,0 @@
-import type { UserSummary } from './user'
-
-export interface AuthSession {
-  accessToken: string | null
-  user: UserSummary | null
-}
